+++ src/helpers/generateTVArticle/GenerateTvArticle.jsx
@@ -1,7 +1,7 @@
 import generateNameTV from "./generateNameTV.js";
 import generatePriceTV from "./generatePriceTV.js";
 import generateStringScreenSizes from "./generateStringScreenSizes.js";
-import GenerateListTvOptions from "../GenerateListTvOptions.jsx";
+import GenerateListTvOptions from "./GenerateListTvOptions.jsx";
 
 import outOfStock from "../../assets/out-of-stock.png";
